feat: add the `recordedtime` extension attribute (#1201)
* feat: add `recordedtime` extension attribute

This extension defines an attribute that represents the time when an occurrence
was recorded in a particular event, i.e. the time when the CloudEvent was
created by a producer.

This enables bitemporal data to be represented with CloudEvents which enables
several use cases described in the specification of the extension attribute.

Signed-off-by: Iwan Aucamp <aucampia@gmail.com>

* changes for review comments

- Mark the attribute as REQUIRED
- Fix a broken sentence
- Removed terminology section as it just repeats what was already
  established earlier.

Signed-off-by: Iwan Aucamp <aucampia@gmail.com>

* Add constraint for relation to occurence time

Given ideal clocks, and given causality holds, occurences should always
be recorded at or after the time of the occurence. But as we don't have
ideal clocks, and might not have causality, it is best to just make this
a loose requirement (i.e. SHOULD) rather than a hard requirement (i.e.
MUST).

Signed-off-by: Iwan Aucamp <aucampia@gmail.com>

---------

Signed-off-by: Iwan Aucamp <aucampia@gmail.com>
Signed-off-by: Alex Collins <alex_collins@intuit.com>